moved the funtions to a new file and created header file
